source/backend/jeevanconnect/requirements.txt
django
djangorestframework
psycopg2
django-cors-headers
channels
daphne
six
geoip2
django-bootstrap5
asgiref